requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
ShopifyAPI==12.3.0
django==5.0.1
psycopg2-binary==2.9.9
djangorestframework==3.14.0
dj-database-url==2.1.0
asgiref==3.7.2
asgiref==3.7.2
django-debug-toolbar==4.2.0
gunicorn==21.2.0 
discord-webhook==1.3.1